chore: move components and fix scroll
